Integrations feature

See merge request spaceuptech/space-cloud-ee!38

+++ gateway/modules/crud/operations.go
@@ -19,6 +19,18 @@ func (m *Module) Create(ctx context.Context, dbAlias, col string, req *model.Cre
 		return err
 	}
 
+	params.Payload = req
+	hookResponse := m.integrationMan.InvokeHook(ctx, params)
+	if hookResponse.CheckResponse() {
+		// Check if an error occurred
+		if err := hookResponse.Error(); err != nil {
+			return err
+		}
+
+		// Gracefully return
+		return nil
+	}
+
 	crud, err := m.getCrudBlock(dbAlias)
 	if err != nil {
 		return err
@@ -72,6 +84,18 @@ func (m *Module) Read(ctx context.Context, dbAlias, col string, req *model.ReadR
 		return nil, err
 	}
 
+	params.Payload = req
+	hookResponse := m.integrationMan.InvokeHook(ctx, params)
+	if hookResponse.CheckResponse() {
+		// Check if an error occurred
+		if err := hookResponse.Error(); err != nil {
+			return nil, err
+		}
+
+		// Gracefully return
+		return hookResponse.Result(), nil
+	}
+
 	if req.IsBatch {
 		key := model.ReadRequestKey{DBType: dbAlias, Col: col, HasOptions: req.Options.HasOptions, Req: *req, ReqParams: params}
 		dataLoader, ok := m.getLoader(fmt.Sprintf("%s-%s-%s", m.project, dbAlias, col))
@@ -104,6 +128,18 @@ func (m *Module) Update(ctx context.Context, dbAlias, col string, req *model.Upd
 
 	if err := m.schema.ValidateUpdateOperation(dbAlias, col, req.Operation, req.Update, req.Find); err != nil {
 		return err
+	}
+
+	params.Payload = req
+	hookResponse := m.integrationMan.InvokeHook(ctx, params)
+	if hookResponse.CheckResponse() {
+		// Check if an error occurred
+		if err := hookResponse.Error(); err != nil {
+			return err
+		}
+
+		// Gracefully return
+		return nil
 	}
 
 	crud, err := m.getCrudBlock(dbAlias)
@@ -158,6 +194,18 @@ func (m *Module) Delete(ctx context.Context, dbAlias, col string, req *model.Del
 		return err
 	}
 
+	params.Payload = req
+	hookResponse := m.integrationMan.InvokeHook(ctx, params)
+	if hookResponse.CheckResponse() {
+		// Check if an error occurred
+		if err := hookResponse.Error(); err != nil {
+			return err
+		}
+
+		// Gracefully return
+		return nil
+	}
+
 	// Invoke the delete intent hook
 	intent, err := m.hooks.Delete(ctx, dbAlias, col, req)
 	if err != nil {
@@ -181,6 +229,19 @@ func (m *Module) Delete(ctx context.Context, dbAlias, col string, req *model.Del
 func (m *Module) ExecPreparedQuery(ctx context.Context, dbAlias, id string, req *model.PreparedQueryRequest, params model.RequestParams) (interface{}, error) {
 	m.RLock()
 	defer m.RUnlock()
+
+	params.Payload = req
+	hookResponse := m.integrationMan.InvokeHook(ctx, params)
+	if hookResponse.CheckResponse() {
+		// Check if an error occurred
+		if err := hookResponse.Error(); err != nil {
+			return nil, err
+		}
+
+		// Gracefully return
+		return hookResponse.Result(), nil
+	}
+
 	crud, err := m.getCrudBlock(dbAlias)
 	if err != nil {
 		return nil, err
@@ -216,6 +277,18 @@ func (m *Module) Aggregate(ctx context.Context, dbAlias, col string, req *model.
 	m.RLock()
 	defer m.RUnlock()
 
+	params.Payload = req
+	hookResponse := m.integrationMan.InvokeHook(ctx, params)
+	if hookResponse.CheckResponse() {
+		// Check if an error occurred
+		if err := hookResponse.Error(); err != nil {
+			return nil, err
+		}
+
+		// Gracefully return
+		return hookResponse.Result(), nil
+	}
+
 	crud, err := m.getCrudBlock(dbAlias)
 	if err != nil {
 		return nil, err
@@ -247,6 +320,18 @@ func (m *Module) Batch(ctx context.Context, dbAlias string, req *model.BatchRequ
 				return err
 			}
 		}
+	}
+
+	params.Payload = req
+	hookResponse := m.integrationMan.InvokeHook(ctx, params)
+	if hookResponse.CheckResponse() {
+		// Check if an error occurred
+		if err := hookResponse.Error(); err != nil {
+			return err
+		}
+
+		// Gracefully return
+		return nil
 	}
 
 	crud, err := m.getCrudBlock(dbAlias)
